renamed chai helpers to correct name
