+++ test/index.js
@@ -1,6 +1,6 @@
 var libContext = require.context("./lib/", true, /Spec.js$/);
 libContext.keys().forEach(libContext);
 
-import FUHelpers from './support/SinonAddons'
+import FUHelpers from './support/ChaiAddons'
 
 chai.use(FUHelpers)
